Categories added
Added the ability to filter the books by their category and also sort on top of the filter. Changed the code for the sql statements to account for this. Added moderate_users.php, will be adding the ability for admins to create and delete users (either admin or regular user).