Admin CRUD for users and some HTML adjustment
Added new edit_user page, to create the section where admins can edit users. Adjusted some pages to have all the new pages at the top of their page. Finished making the moderate_users page so it outputs a table of all the users for an admin to edit, as well as a create user section at the top.